+++ web/hooks/command.ts
@@ -14,7 +14,7 @@ export type CommandDef = {
 
 export type AppCommand = {
   command: CommandDef;
-  run: (args?: string[]) => Promise<void>;
+  run: (args?: any[]) => Promise<void>;
 };
 
 export type CommandHookT = {
@@ -44,7 +44,7 @@ export class CommandHook extends EventEmitter<CommandHookEvents>
         this.editorCommands.set(cmd.name, {
           command: cmd,
           run: (args?: string[]) => {
-            return plug.invoke(name, [cmd, ...args??[]]);
+            return plug.invoke(name, [cmd, ...args ?? []]);
           },
         });
       }