+++ website/CHANGELOG.md
@@ -3,6 +3,11 @@ release.
 
 ---
 
+## Next
+* Keyboard shortcuts can now be configured in [[SETTINGS]]
+
+---
+
 ## 0.5.10
 * **Breaking change**: Local attachment URLs (`[page](url)` syntax and `![alt](url)` image syntax) are now interpreted to relative to the page's folder, unless their URL starts with a `/` then they're relative to the space root (as per [this issue](https://github.com/silverbulletmd/silverbullet/issues/363))
 * **Breaking change:** Revamped [[Templates]], specifically changed the format of [[Page Templates]]. The “Template: Instantiate Page” has been renamed to {[Page: From Template]}.